Extract CSS tests
