+++ packages/wmr/test/fixtures.test.js
@@ -11,19 +11,14 @@ import {
 	waitForNotMessage,
 	waitFor,
 	withLog,
-	waitForPass
+	waitForPass,
+	updateFile
 } from './test-helpers.js';
 import { rollup } from 'rollup';
 import nodeBuiltinsPlugin from '../src/plugins/node-builtins-plugin.js';
 import { supportsSearchParams } from '../src/lib/net-utils.js';
 
 jest.setTimeout(30000);
-
-async function updateFile(tempDir, file, replacer) {
-	const compPath = path.join(tempDir, file);
-	const content = await fs.readFile(compPath, 'utf-8');
-	await fs.writeFile(compPath, replacer(content));
-}
 
 describe('fixtures', () => {
 	/** @type {TestEnv} */
@@ -298,16 +293,16 @@ describe('fixtures', () => {
 			await withLog(instance.output, async () => {
 				await getOutput(env, instance);
 
-				await await waitForPass(async () => {
+				await waitForPass(async () => {
 					const color = await env.page.$eval('h1', el => getComputedStyle(el).color);
 					expect(color).toBe('rgb(255, 218, 185)');
 				});
 
-				updateFile(env.tmp.path, 'foo/style.css', () => {
+				await updateFile(env.tmp.path, 'foo/style.css', () => {
 					return `h1 { color: red; }`;
 				});
 
-				await await waitForPass(async () => {
+				await waitForPass(async () => {
 					const color = await env.page.$eval('h1', el => getComputedStyle(el).color);
 					expect(color).toBe('rgb(255, 0, 0)');
 				});
@@ -360,40 +355,6 @@ describe('fixtures', () => {
 			await getOutput(env, instance);
 			expect(await env.page.$eval('#external', el => el.textContent)).toBe('rendered from unpkg');
 			expect(await env.page.evaluate(`window.unistore`)).toBeTruthy();
-		});
-	});
-
-	describe('CSS', () => {
-		it('should load referenced files via @import', async () => {
-			await loadFixture('css-imports', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			await withLog(instance.output, async () => {
-				expect(await env.page.$eval('h1', el => getComputedStyle(el).color)).toBe('rgb(255, 0, 0)');
-			});
-		});
-
-		it('should load referenced files via url()', async () => {
-			await loadFixture('css-imports', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-			expect(await env.page.$eval('body', el => getComputedStyle(el).backgroundImage)).toMatch(/img\.jpg/);
-		});
-
-		it('should warn on CSS modules with reserved class names', async () => {
-			await loadFixture('css-module-reserved', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			await waitForPass(async () => {
-				expect(await env.page.$eval('.foo', el => getComputedStyle(el).backgroundColor)).toBe('rgb(255, 218, 185)');
-				expect(await env.page.$eval('.new', el => getComputedStyle(el).backgroundColor)).toBe('rgb(255, 255, 0)');
-				expect(await env.page.$eval('.debugger', el => getComputedStyle(el).backgroundColor)).toBe('rgb(255, 0, 0)');
-				expect(await env.page.$eval('.const', el => getComputedStyle(el).backgroundColor)).toBe('rgb(255, 218, 185)');
-			});
-
-			expect(instance.output.join('\n')).toMatch(/Cannot use reserved word/);
 		});
 	});
 
@@ -551,86 +512,6 @@ describe('fixtures', () => {
 			const child = await env.page.$('.child');
 			text = child ? await child.evaluate(el => el.textContent) : null;
 			expect(text).toEqual('child');
-		});
-
-		it('should hot reload a css-file imported from index.html', async () => {
-			await loadFixture('hmr', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			expect(await page.$eval('body', e => getComputedStyle(e).color)).toBe('rgb(51, 51, 51)');
-
-			await updateFile(env.tmp.path, 'index.css', content => content.replace('color: #333;', 'color: #000;'));
-
-			await waitForPass(async () => {
-				expect(await page.$eval('body', e => getComputedStyle(e).color)).toBe('rgb(0, 0, 0)');
-			});
-		});
-
-		it('should hot reload a module css-file', async () => {
-			await loadFixture('hmr', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			expect(await page.$eval('main', e => getComputedStyle(e).color)).toBe('rgb(51, 51, 51)');
-
-			await updateFile(env.tmp.path, 'style.module.css', content => content.replace('color: #333;', 'color: #000;'));
-
-			await waitForPass(async () => {
-				expect(await page.$eval('main', e => getComputedStyle(e).color)).toBe('rgb(0, 0, 0)');
-			});
-		});
-
-		it('should hot reload a css-file', async () => {
-			await loadFixture('hmr-css', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			await withLog(instance.output, async () => {
-				expect(await page.$eval('h1', e => getComputedStyle(e).color)).toBe('rgb(51, 51, 51)');
-
-				await updateFile(env.tmp.path, path.join('public', 'index.css'), content =>
-					content.replace('background: white;', 'background: red;')
-				);
-
-				await waitForPass(async () => {
-					expect(await page.$eval('body', e => getComputedStyle(e).backgroundColor)).toBe('rgb(255, 0, 0)');
-				});
-			});
-		});
-
-		it('should hot reload a nested css-file', async () => {
-			await loadFixture('hmr-css', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			await withLog(instance.output, async () => {
-				expect(await page.$eval('h1', e => getComputedStyle(e).color)).toBe('rgb(51, 51, 51)');
-
-				await updateFile(env.tmp.path, path.join('public', 'home.css'), content =>
-					content.replace('color: #333;', 'color: red;')
-				);
-
-				await waitForPass(async () => {
-					expect(await page.$eval('h1', e => getComputedStyle(e).color)).toBe('rgb(255, 0, 0)');
-				});
-			});
-		});
-
-		it('should hot reload a nested css-file with no public folder', async () => {
-			await loadFixture('hmr-css-no-public', env);
-			instance = await runWmrFast(env.tmp.path);
-			await getOutput(env, instance);
-
-			await withLog(instance.output, async () => {
-				expect(await page.$eval('h1', e => getComputedStyle(e).color)).toBe('rgb(51, 51, 51)');
-
-				await updateFile(env.tmp.path, 'home.css', content => content.replace('color: #333;', 'color: red;'));
-
-				await waitForPass(async () => {
-					expect(await page.$eval('h1', e => getComputedStyle(e).color)).toBe('rgb(255, 0, 0)');
-				});
-			});
 		});
 	});
